Add the possibility to clear the NFVO after each test from remainings of the last test
If a test failed before this commit, the following tests would most likely also fail because the integration test
could not upload the VIM or NSD due to the remaining ones from the previous test.
Now one may specify the option to clear the NFVO after each test in the properties.
Additionaly at the end of the integration tests a line will now show up if there were failing tests.