Revise the scaling scenario
The scaling scenario using iperf was not very reliable and it was hard to figure out if an error in this test case originated in a problem with Openbaton or in a timed out iperf connection. So the new scaling scenario test uses ncat to transmit a message from the clients to the servers so that the integration test does not have to check for the presence of a connection but just if the message was received by the servers.